tema 2

+++ Ejercicio23t2/src/ejercicio23t2/Ejercicio23t2.java
@@ -15,14 +15,17 @@ public class Ejercicio23t2 {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        double precio;
+        float precio;
         int uds;
         Scanner teclado = new Scanner(System.in);
+        
         System.out.println("Por favor, introduzca el precio del modelo de ordenador que desea comprar: ");
-        precio = teclado.nextDouble();
+        precio = teclado.nextFloat();
+        
         System.out.println("¿Cuántas unidades quiere llevarse? ");
         uds=teclado.nextInt();
-        double resultado= precio*uds;
+        
+        float resultado= precio*uds;
         System.out.println("El precio total de su compra es: "+resultado);
     }
